# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Ce {
  id: number;
  slug: string;
  name: string;
  city: string;
  country: string;
  category: string;
  summary: string;
  emoji: string;
  status: string;
  chartCount: number;
  createdAt: string;
  updatedAt: string;
}

export type ChartSpec = { [key: string]: unknown };

export type ChartProvenance = { [key: string]: unknown } | null;

export interface Chart {
  id: number;
  ceId: number;
  slug: string;
  question: string;
  title: string;
  subtitle: string;
  insight: string;
  chartType: string;
  spec: ChartSpec;
  status?: string;
  provenance?: ChartProvenance;
  sortOrder: number;
  openFeedbackCount?: number;
  topFeedbackSeverity?: string | null;
  editCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface ChartWithCe {
  chart: Chart;
  ce: Ce;
}

export interface CeWithCharts {
  ce: Ce;
  charts: Chart[];
}

export type DrdSourcesItem = { [key: string]: unknown };

export interface Drd {
  id: number;
  ceSlug: string;
  sourceType: string;
  sourceFilename?: string | null;
  sources: DrdSourcesItem[];
  markdownPreview: string;
  markdownLength: number;
  createdAt: string;
  updatedAt: string;
}

export type DrdMarkdownUploadSourcesItem = { [key: string]: unknown };

export interface DrdMarkdownUpload {
  ceSlug: string;
  markdown: string;
  sourceFilename?: string;
  sources?: DrdMarkdownUploadSourcesItem[];
}

export interface Subcategory {
  id: string;
  label: string;
  description: string;
}

export interface ResearchGenerateInput {
  ceSlug: string;
  subcategoryId: string;
  subcategoryLabel?: string;
  subcategoryDescription?: string;
  name?: string;
  city?: string;
  country?: string;
  category?: string;
  writerTopics?: string[];
}

export type ResearchGenerateResultDroppedQuestionsItem = {
  question: string;
  reason: string;
};

export type ResearchGenerateResultProposedHeroQuestionsItem = {
  [key: string]: unknown;
};

export interface ResearchGenerateResult {
  ce: Ce;
  charts: Chart[];
  publishedChartsKept: number;
  droppedQuestions: ResearchGenerateResultDroppedQuestionsItem[];
  proposedHeroQuestions: ResearchGenerateResultProposedHeroQuestionsItem[];
}

export interface ChartFeedback {
  id: number;
  chartId: number;
  rating?: number | null;
  issueCategory?: string | null;
  note: string;
  reporterName: string;
  status: string;
  severity: string;
  createdAt: string;
  updatedAt: string;
}

export interface ChartFeedbackInput {
  rating?: number | null;
  issueCategory?: string | null;
  note?: string;
  reporterName?: string;
}

export interface ChartEditInput {
  title?: string;
  subtitle?: string;
  insight?: string;
  editorName?: string;
  summary?: string;
}

export interface FeedbackUpdate {
  status: string;
}

export interface TriageFeedback {
  feedback: ChartFeedback;
  chart: Chart;
  ce: Ce;
  chartEditCount: number;
}

export interface QuestionTroubleScore {
  questionText: string;
  subcategoryId: string;
  chartType: string;
  chartCount: number;
  openFeedbackCount: number;
  editCount: number;
  troubleScore: number;
  sampleCeSlugs: string[];
}

export interface CeInput {
  name: string;
  city: string;
  country: string;
  category?: string;
}

export type EditChart200 = {
  chart: Chart;
};

export type ListAllFeedbackParams = {
  status?: string;
};