# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ces:
    get:
      operationId: listCes
      tags: [ces]
      summary: List all CEs
      responses:
        "200":
          description: List of CEs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ce"
    post:
      operationId: createCe
      tags: [ces]
      summary: Create a CE and generate its chart suite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CeWithCharts"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /ces/{slug}:
    get:
      operationId: getCe
      tags: [ces]
      summary: Get a CE with its charts
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CE with charts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CeWithCharts"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteCe
      tags: [ces]
      summary: Delete a CE and its charts
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /ces/{slug}/regenerate:
    post:
      operationId: regenerateCe
      tags: [ces]
      summary: Replace this CE's charts with a freshly generated set
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Regenerated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CeWithCharts"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: AI generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /charts/{id}:
    get:
      operationId: getChart
      tags: [charts]
      summary: Get a single chart
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChartWithCe"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: editChart
      tags: [charts]
      summary: Edit a chart's writer-controlled copy
      description: |
        Updates title/subtitle/insight and logs a chart_edits row. The edit
        row counts as a 0.25-weight implicit signal in the trouble score.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChartEditInput"
      responses:
        "200":
          description: Updated chart
          content:
            application/json:
              schema:
                type: object
                required: [chart]
                properties:
                  chart:
                    $ref: "#/components/schemas/Chart"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Ce:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        city:
          type: string
        country:
          type: string
        category:
          type: string
        summary:
          type: string
        emoji:
          type: string
        status:
          type: string
        chartCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - name
        - city
        - country
        - category
        - summary
        - emoji
        - status
        - chartCount
        - createdAt
        - updatedAt
    Chart:
      type: object
      properties:
        id:
          type: integer
        ceId:
          type: integer
        slug:
          type: string
        question:
          type: string
        title:
          type: string
        subtitle:
          type: string
        insight:
          type: string
        chartType:
          type: string
        spec:
          type: object
          additionalProperties: true
        status:
          type: string
        provenance:
          type: object
          nullable: true
          additionalProperties: true
        sortOrder:
          type: integer
        openFeedbackCount:
          type: integer
        topFeedbackSeverity:
          type: string
          nullable: true
        editCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - ceId
        - slug
        - question
        - title
        - subtitle
        - insight
        - chartType
        - spec
        - sortOrder
        - createdAt
        - updatedAt
    ChartWithCe:
      type: object
      properties:
        chart:
          $ref: "#/components/schemas/Chart"
        ce:
          $ref: "#/components/schemas/Ce"
      required:
        - chart
        - ce
    CeWithCharts:
      type: object
      properties:
        ce:
          $ref: "#/components/schemas/Ce"
        charts:
          type: array
          items:
            $ref: "#/components/schemas/Chart"
      required:
        - ce
        - charts
...
    ChartEditInput:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        insight:
          type: string
        editorName:
          type: string
        summary:
          type: string
...
    CeInput:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        country:
          type: string
        category:
          type: string
      required:
        - name
        - city
        - country