# lib/api-zod/src/generated/types/chart.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChartProvenance } from "./chartProvenance";
import type { ChartSpec } from "./chartSpec";

export interface Chart {
  id: number;
  ceId: number;
  slug: string;
  question: string;
  title: string;
  subtitle: string;
  insight: string;
  chartType: string;
  spec: ChartSpec;
  status?: string;
  provenance?: ChartProvenance;
  sortOrder: number;
  openFeedbackCount?: number;
  topFeedbackSeverity?: string | null;
  editCount?: number;
  createdAt: string;
  updatedAt: string;
}