# lib/api-zod/src/generated/types/chartEditInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChartEditInput {
  title?: string;
  subtitle?: string;
  insight?: string;
  editorName?: string;
  summary?: string;
}